Move hasFindMusicPermissions to utilities
Moved hasFindMusicPermissions from clients/findmusic.ts to utilities.ts
as it's a browser extension permission check, not an API client function.
Updated all imports and tests accordingly.
