+++ test/findmusic.test.ts
@@ -11,7 +11,8 @@ vi.mock('../src/logger', () => ({
 
 vi.mock('../src/utilities', () => ({
   storeFindMusicToken: vi.fn(),
-  getFindMusicTokenFromStorage: vi.fn()
+  getFindMusicTokenFromStorage: vi.fn(),
+  hasFindMusicPermissions: vi.fn()
 }));
 
 const mockCookiesGet = vi.fn();
@@ -29,7 +30,7 @@ global.fetch = vi.fn();
 process.env.FINDMUSIC_BASE_URL = 'https://findmusic.club';
 
 import { exchangeBandcampToken, getFindMusicToken } from '../src/clients/findmusic';
-import { storeFindMusicToken, getFindMusicTokenFromStorage } from '../src/utilities';
+import { storeFindMusicToken, getFindMusicTokenFromStorage, hasFindMusicPermissions } from '../src/utilities';
 
 describe('FindMusic Client', () => {
   beforeEach(() => {
@@ -184,7 +185,9 @@ describe('FindMusic Client', () => {
 
     it('should return stored token if valid', async () => {
       const mockStoredToken = 'stored-jwt-token';
+      const hasFindMusicPermissionsMock = vi.mocked(hasFindMusicPermissions);
       const getFindMusicTokenFromStorageMock = vi.mocked(getFindMusicTokenFromStorage);
+      hasFindMusicPermissionsMock.mockResolvedValue(true);
       getFindMusicTokenFromStorageMock.mockResolvedValue(mockStoredToken);
 
       const result = await getFindMusicToken();
@@ -196,9 +199,11 @@ describe('FindMusic Client', () => {
 
     it('should exchange new token if no stored token', async () => {
       const mockJwtToken = 'new-jwt-token';
+      const hasFindMusicPermissionsMock = vi.mocked(hasFindMusicPermissions);
       const getFindMusicTokenFromStorageMock = vi.mocked(getFindMusicTokenFromStorage);
       const storeFindMusicTokenMock = vi.mocked(storeFindMusicToken);
 
+      hasFindMusicPermissionsMock.mockResolvedValue(true);
       getFindMusicTokenFromStorageMock.mockResolvedValue(null);
       mockCookiesGet.mockResolvedValue({ value: 'test-bc-token' });
       vi.mocked(global.fetch).mockResolvedValue({
